+++ App/xcbuddykit/Sources/Generator/WorkspaceGenerator.swift
@@ -1,7 +1,11 @@
 import Basic
 import Foundation
+import PathKit
+import xcproj
 
 protocol WorkspaceGenerating: AnyObject {
+    func generate(path: AbsolutePath,
+                  generatorContext: GeneratorContexting) throws
 }
 
 final class WorkspaceGenerator: WorkspaceGenerating {
@@ -11,6 +15,12 @@ final class WorkspaceGenerator: WorkspaceGenerating {
         self.projectGenerator = projectGenerator
     }
 
-    func generate(path _: AbsolutePath) {
+    func generate(path: AbsolutePath,
+                  generatorContext _: GeneratorContexting) throws {
+        let workspacePath = Path(path.appending(component: Constants.Xcode.workspaceName).asString)
+        let workspaceData = XCWorkspaceData(children: [])
+        let workspace = XCWorkspace(data: workspaceData)
+
+        try workspace.write(path: workspacePath, override: true)
     }
 }