+++ App/xcbuddykit/Sources/Loader/GraphLoaderContext.swift
@@ -11,6 +11,9 @@ protocol GraphLoaderContexting: AnyObject {
 
     /// Util to handle files.
     var fileHandler: FileHandling { get }
+
+    /// Circular dependency detector.
+    var circularDetector: GraphCircularDetecting { get }
 }
 
 /// Object passed during the graph loading that contains utils to be used.
@@ -24,24 +27,23 @@ class GraphLoaderContext: GraphLoaderContexting {
     /// Util to handle files.
     let fileHandler: FileHandling
 
+    /// Circular dependency detector.
+    let circularDetector: GraphCircularDetecting
+
     /// Initializes the context with its attributes.
     ///
     /// - Parameters:
     ///   - manifestLoader: Manifest loader that is used to get a JSON representation of the manifests.
     ///   - cache: Contains a reference to the manifests that are parsed during the graph loading.
     ///   - fileHandler: Util to handle files.
-    init(manifestLoader: GraphManifestLoading,
-         cache: GraphLoaderCaching,
-         fileHandler: FileHandling = FileHandler()) {
+    ///   - circularDetector: Circular dependency detector.
+    init(manifestLoader: GraphManifestLoading = GraphManifestLoader(),
+         cache: GraphLoaderCaching = GraphLoaderCache(),
+         fileHandler: FileHandling = FileHandler(),
+         circularDetector: GraphCircularDetecting = GraphCircularDetector()) {
         self.manifestLoader = manifestLoader
         self.cache = cache
         self.fileHandler = fileHandler
-    }
-
-    /// Default context constructor.
-    init() {
-        manifestLoader = GraphManifestLoader()
-        cache = GraphLoaderCache()
-        fileHandler = FileHandler()
+        self.circularDetector = circularDetector
     }
 }