+++ App/xcbuddykit/Tests/Generator/MockProjectGenerator.swift
@@ -1,0 +1,5 @@
+import Foundation
+@testable import xcbuddykit
+
+final class MockProjectGenerator: ProjectGenerating {
+}