+++ App/xcbuddykit/Tests/Loader/MockGraphLoaderCache.swift
@@ -3,6 +3,9 @@ import Foundation
 @testable import xcbuddykit
 
 final class MockGraphLoaderCache: GraphLoaderCaching {
+    var projects: [AbsolutePath: Project] = [:]
+    var targetNodes: [AbsolutePath: [String: TargetNode]] = [:]
+    var precompiledNodes: [AbsolutePath: PrecompiledNode] = [:]
     var projectStub: ((AbsolutePath) -> Project?)?
     var projectCount: UInt = 0
     var addProjectCount: UInt = 0
@@ -11,10 +14,14 @@ final class MockGraphLoaderCache: GraphLoaderCaching {
     var addConfigArgs: [Config] = []
     var configStub: ((AbsolutePath) -> Config?)?
     var configCount: UInt = 0
-    var addNodeCount: UInt = 0
-    var addNodeArgs: [GraphNode] = []
-    var nodeCount: UInt = 0
-    var nodeStub: ((AbsolutePath) -> GraphNode?)?
+    var addPrecompiledNodeCount: UInt = 0
+    var addPrecompiledArgs: [PrecompiledNode] = []
+    var precompiledNodeCount: UInt = 0
+    var precompiledNodeStub: ((AbsolutePath) -> PrecompiledNode?)?
+    var addTargetNodeCount: UInt = 0
+    var addTargetNodeArgs: [TargetNode] = []
+    var targetNodeCount: UInt = 0
+    var targetNodeStub: ((AbsolutePath, String) -> TargetNode?)?
 
     func project(_ path: AbsolutePath) -> Project? {
         projectCount += 1
@@ -36,13 +43,23 @@ final class MockGraphLoaderCache: GraphLoaderCaching {
         addConfigArgs.append(config)
     }
 
-    func add(node: GraphNode) {
-        addNodeCount += 1
-        addNodeArgs.append(node)
+    func add(precompiledNode: PrecompiledNode) {
+        addPrecompiledNodeCount += 1
+        addPrecompiledArgs.append(precompiledNode)
     }
 
-    func node(_ path: AbsolutePath) -> GraphNode? {
-        nodeCount += 1
-        return nodeStub?(path)
+    func precompiledNode(_ path: AbsolutePath) -> PrecompiledNode? {
+        precompiledNodeCount += 1
+        return precompiledNodeStub?(path)
+    }
+
+    func add(targetNode: TargetNode) {
+        addTargetNodeCount += 1
+        addTargetNodeArgs.append(targetNode)
+    }
+
+    func targetNode(_ path: AbsolutePath, name: String) -> TargetNode? {
+        targetNodeCount += 1
+        return targetNodeStub?(path, name)
     }
 }